Upload code with gates from Brewster paper
Made a tweak because realized I misinterpreted some of the gates they want implemented but still do not get quantum results. I don't think I'm correctly understanding how Brewster intends for the circuit to be implemented. It seems there are 3 parameters that have to be tuned; I don't know how to tune those in this context. More clarification is needed.